Handle zotero by ensuring application is closed before backup.
Added name attr to namedtuple for convenience with zotero

Somewhat tricky, as relying on pgrep ans os.kill, so a few try-except
blocks that should be wrapped up into a nicer structure ... but hey.

pgrep has a status code of 1 for not found ... handled by except.

separately, the kill process is simply tried and excepted without
handling.